move role-validator tests next to the validator
TestValidateModelTaskAssignment and its _task_validation_enabled fixture
were still living in tests/test_config.py even though the code under
test now lives in lilbee.modelhub.role_validator. Move both to a new
tests/test_role_validator.py so the tests are next to what they cover.

Also drop the unused _validator_env helper that's been dead since the
fixture replaced it.
